requirements.txt
starlite==1.51.0
deta==1.1.0
python-socketio==5.7.2
python-engineio==4.3.4
websockets==10.4
python-dotenv==0.21.1
odetam==1.2.0
aiodeta==0.1.2
bcrypt==4.0.1
python-jose==3.3.0
Jinja2==3.1.2
python-magic==0.4.27






